bot/requirements.txt
Flask
Flask-CORS
python-dotenv
numpy
google-generativeai
firebase-admin
python-docx
langchain
